udpates to server.go and cleaned up main.go code
